add linux support and fix a few bugs

+++ src/cmdlib.h
@@ -54,7 +54,7 @@ extern char **myargv;
 
 //char *strupr(char *in);
 //char *strlower(char *in);
-int Q_strncasecmp(const char *s1, const char *s2, int n);
+int Q_strncasecmp(const char *s1, const char *s2, size_t n);
 int Q_strcasecmp(const char *s1, const char *s2);
 void Q_getwd(char *out);
